UOELSA-1102: Korjaa tietotyyppi
- Muuta erikoistumiseen vaadittava tuntimäärä kokonaisluvusta liukuluvuksi vastaamaan opintooppaalle määritettyä tyyppiä

+++ src/main/kotlin/fi/elsapalvelu/elsa/service/dto/TeoriakoulutusDTO.kt
@@ -22,7 +22,7 @@ data class TeoriakoulutusDTO(
     var paattymispaiva: LocalDate? = null,
 
     @get: Min(value = 0)
-    var erikoistumiseenHyvaksyttavaTuntimaara: Int? = null,
+    var erikoistumiseenHyvaksyttavaTuntimaara: Double? = null,
 
     var todistukset: MutableSet<AsiakirjaDTO>? = mutableSetOf()
